Final_0403_Extracted/extract_0329G/requirements.txt
# ✅ Core Dependencies
openai==1.2.3
requests==2.31.0
pandas==2.0.3
yfinance==0.2.31
python-dotenv==1.0.0

# ✅ Trading API Dependencies
binance-connector==3.1.0
#ig-markets-api==0.2.1
alpha_vantage==2.3.1

# ✅ Additional Tools
pytest==7.4.3
matplotlib==3.7.2

